chage integer

+++ app/models.py
@@ -9,7 +9,7 @@ class Debtor(models.Model):
     address = models.TextField(blank=True, null= True)
     img = models.ImageField(upload_to='pics', null=True, blank=True)
     lender = models.ForeignKey(User, on_delete=models.SET_NULL, blank=True, null=True)
-    totalbalance = models.IntegerField(default=0, null=True, blank=True)
+    totalbalance = models.FloatField(default=0, null=True, blank=True)
 
     def __str__(self):
         return "PA00" +str(self.id)
@@ -17,9 +17,9 @@ class Debtor(models.Model):
 class Account(models.Model):
     debtor = models.ForeignKey(Debtor, on_delete=models.SET_NULL, blank=True, null=True)
     description = models.TextField(blank=True, null= True)
-    loan = models.IntegerField(default=0, null=True, blank=True)
-    deposit = models.IntegerField(default=0, null=True, blank=True)
-    balance = models.IntegerField(default=0, null=True, blank=True)
+    loan = models.FloatField(default=0, null=True, blank=True)
+    deposit = models.FloatField(default=0, null=True, blank=True)
+    balance = models.FloatField(default=0, null=True, blank=True)
     date = models.DateTimeField(auto_now_add=True)
     status = models.BooleanField(default=False, null=True, blank=False)    
 
@@ -29,7 +29,7 @@ class Lender(models.Model):
     address = models.TextField(blank=True, null= True)
     img = models.ImageField(upload_to='pics', null=True, blank=True)
     investor = models.ForeignKey(User, on_delete=models.SET_NULL, blank=True, null=True)
-    totalbalance = models.IntegerField(default=0, null=True, blank=True)    
+    totalbalance = models.FloatField(default=0, null=True, blank=True)    
     
     def __str__(self):
         return "IV00" +str(self.id)
@@ -37,9 +37,9 @@ class Lender(models.Model):
 class Invest(models.Model):
     lender = models.ForeignKey(Lender, on_delete=models.SET_NULL, blank=True, null=True)
     description = models.TextField(blank=True, null= True)
-    invest = models.IntegerField(default=0, null=True, blank=True)
-    retern = models.IntegerField(default=0, null=True, blank=True)
-    balance = models.IntegerField(default=0, null=True, blank=True)
+    invest = models.FloatField(default=0, null=True, blank=True)
+    retern = models.FloatField(default=0, null=True, blank=True)
+    balance = models.FloatField(default=0, null=True, blank=True)
     date = models.DateTimeField(auto_now_add=True)
     status = models.BooleanField(default=False, null=True, blank=False)        
 
@@ -53,9 +53,9 @@ class Landowner(models.Model):
 
 class Land(models.Model):
     landowner = models.ForeignKey(Landowner, on_delete=models.SET_NULL, blank=True, null=True)
-    landQTY = models.IntegerField(default=0, null=True, blank=True)
+    landQTY = models.FloatField(default=0, null=True, blank=True)
     rsdag = models.CharField(max_length=50)
-    perDprice = models.IntegerField(default=0, null=True, blank=True)
+    perDprice = models.FloatField(default=0, null=True, blank=True)
     mouja = models.CharField(max_length=50)
 
     def totalprice(self):
@@ -65,4 +65,4 @@ class Advance(models.Model):
     land = models.ForeignKey(Land, on_delete=models.SET_NULL, blank=True, null=True)
     date = models.DateTimeField(auto_now_add=True)
     description = models.TextField(blank=True, null= True)
-    advance = models.IntegerField(default=0, null=True, blank=True)
+    advance = models.FloatField(default=0, null=True, blank=True)
